+++ src/main/java/com/intervale/cources/library/SpringBookDemoApplication.java
@@ -1,0 +1,12 @@
+package com.intervale.cources.library;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpringBookDemoApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpringBookDemoApplication.class, args);
+    }
+}
