+++ src/main/java/com/intervale/cources/library/config/BeansConfig.java
@@ -1,0 +1,19 @@
+package com.intervale.cources.library.config;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.web.client.RestTemplateBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class BeansConfig {
+    @Bean
+    public RestTemplate restTemplate(RestTemplateBuilder builder) {
+        return builder.build();
+    }
+    @Bean
+    public ModelMapper getMapperModel() {
+        return new ModelMapper();
+    }
+}
