+++ src/main/java/com/intervale/cources/library/openlibrary/api/OpenLibraryController.java
@@ -1,0 +1,9 @@
+package com.intervale.cources.library.openlibrary.api;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.RestController;
+
+
+public class OpenLibraryController {
+
+}
