--- src/main/java/intervale/dz3/springbookdemo/SpringBookDemoApplication.java
@@ -1,18 +1,0 @@
-package intervale.dz3.springbookdemo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DriverManagerDataSource;
-import org.springframework.web.client.RestTemplate;
-
-import javax.sql.DataSource;
-
-@SpringBootApplication
-public class SpringBookDemoApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpringBookDemoApplication.class, args);
-    }
-}
