--- src/main/java/intervale/dz3/springbookdemo/config/BeansConfig.java
@@ -1,19 +1,0 @@
-package intervale.dz3.springbookdemo.config;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class BeansConfig {
-    @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder builder) {
-        return builder.build();
-    }
-    @Bean
-    public ModelMapper getMapperModel() {
-        return new ModelMapper();
-    }
-}
