--- src/main/java/intervale/dz3/springbookdemo/model/BooksDto.java
@@ -1,31 +1,0 @@
-package intervale.dz3.springbookdemo.model;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BooksDto {
-    private Integer id;
-
-    private String isbn;
-
-    private String name;
-
-    private String author;
-
-    private int pages;
-
-    private double weight;
-
-    private int price;
-
-//    @JsonProperty("name")
-//    BooksDto userDto;
-}
