--- src/main/java/intervale/dz3/springbookdemo/openlibrary/OpenLibrarySearchResponse.java
@@ -1,19 +1,0 @@
-package intervale.dz3.springbookdemo.openlibrary;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-public class OpenLibrarySearchResponse {
-
-    private Integer start;
-
-    @JsonProperty("num_found")
-    private Integer numFound;
-
-    private List<OpenLibraryDocs> docs;
-
-
-}
