--- src/main/java/intervale/dz3/springbookdemo/openlibrary/api/OpenLibraryController.java
@@ -1,10 +1,0 @@
-package intervale.dz3.springbookdemo.openlibrary.api;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-public class OpenLibraryController {
-
-}
