+++ src/test/java/com/intervale/cources/SpringBookDemoApplicationTests.java
@@ -1,0 +1,21 @@
+package com.intervale.cources;
+
+
+import org.junit.Test;
+
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import static org.mockito.Mockito.mock;
+
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringBootTest
+class SpringBookDemoApplicationTests {
+
+    @Test
+    void contextLoads() {
+
+    }
+
+}
