--- src/test/java/intervale/dz3/springbookdemo/SpringBookDemoApplicationTests.java
@@ -1,25 +1,0 @@
-package intervale.dz3.springbookdemo;
-
-
-import intervale.dz3.springbookdemo.model.Books;
-import org.junit.Test;
-
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-import java.sql.ResultSet;
-
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringBootTest
-class SpringBookDemoApplicationTests {
-
-    @Test
-    void contextLoads() {
-
-    }
-
-}
